backend/pyproject.toml
[project]
name = "FastAPI-backend"
version = "0.1.0"
description = "FastAPI application with Gemini and FireworksAI"
requires-python = ">=3.12"
authors = [
    { name = "Leo Walker" }
]
dependencies = [
    "pydantic",
    "typing-extensions",
    "ipykernel",
    "logfire>=3.5.3",
    "uvicorn[standard]>=0.34.0",
    "gunicorn>=23.0.0",
    "litellm>=1.60.6",
    "fastapi>=0.115.8",
    "python-dotenv>=1.0.1",
    "sse-starlette>=2.2.1",
    "hatchling>=1.27.0",
]

# Add development dependencies section
[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "mypy"
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
